Move MOSTLY_STABLE_DELAY const into src folder
