+++ src/components/SplitHorizontal/SplitHorizontal.spec.tsx
@@ -10,7 +10,7 @@ import SplitHorizontal, {
 } from './SplitHorizontal';
 import DragCaptureZone from '../DragCaptureZone/DragCaptureZone';
 import { Motion } from 'react-motion';
-import { MOSTLY_STABLE_DELAY } from '../../../tests/constants';
+import { MOSTLY_STABLE_DELAY } from '../../util/constants';
 
 describe('SplitHorizontal', () => {
 	common(SplitHorizontal);
